skal hente dekoratør på annen måte. Fjerner request og mustache-express
