+++ package.json
@@ -9,6 +9,7 @@
     "@navikt/ds-icons": "^2.8.1",
     "@navikt/ds-react": "^2.8.1",
     "@navikt/ds-tokens": "^2.8.1",
+    "@navikt/nav-dekoratoren-moduler": "^1.9.0",
     "@sanity/client": "^3.4.1",
     "@types/amplitude-js": "^8.16.2",
     "@types/marked": "^4.0.8",
@@ -20,12 +21,10 @@
     "express": "^4.18.2",
     "jsdom": "^21.1.1",
     "marked": "^4.2.12",
-    "mustache-express": "^1.3.2",
     "react": "^18.2.0",
     "react-app-polyfill": "^3.0.0",
     "react-dom": "^18.2.0",
     "react-scripts": "^5.0.1",
-    "request": "^2.88.2",
     "smoothscroll-polyfill": "^0.4.4",
     "styled-components": "^5.3.9",
     "typescript": "^4.9.4"
@@ -37,8 +36,8 @@
     "eject": "react-scripts eject",
     "predeploy": "craco build",
     "snyk-protect": "snyk protect",
-    "serve": "NODE_ENV=production node server.js",
-    "serve:dev": "NODE_ENV=local node server.js"
+    "serve": "NODE_ENV=production node src/backend/server.js",
+    "serve:dev": "NODE_ENV=local node src/backend/server.js"
   },
   "browserslist": {
     "production": [
