+++ public/index.html
@@ -12,34 +12,15 @@
 		/>
 		<meta http-equiv="Pragma" content="no-cache" />
 		<meta name="theme-color" content="#000000" />
-
-		<!-- Hotjar Tracking Code for http://navlab.no/ -->
-		<script>
-			(function(h,o,t,j,a,r){
-        		h.hj=h.hj||function(){(h.hj.q=h.hj.q||[]).push(arguments)};
-        		h._hjSettings={hjid:118350,hjsv:6};
-        		a=o.getElementsByTagName('head')[0];
-        		r=o.createElement('script');r.async=1;
-        		r.src=t+h._hjSettings.hjid+j+h._hjSettings.hjsv;
-        		a.appendChild(r);
-      		})(window,document,'https://static.hotjar.com/c/hotjar-','.js?sv=');
-    	</script>
-
 		<link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
 		<link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
 		<link rel="stylesheet" href="%PUBLIC_URL%/index.css" />
-
-    {{{STYLES}}}
-
 		<title>Enslig mor eller far - www.nav.no</title>
 	</head>
 	<body>
 		<noscript> You need to enable JavaScript to run this app. </noscript>
 		<div class="pagewrapper">
-      		{{{HEADER}}}
 			<div id="root" role="document"></div>
 		</div>
-    	{{{SCRIPTS}}}
 	</body>
-  	{{{FOOTER}}}
 </html>
